requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
langgraph
langchain-google-genai
langchain-core
pydantic>=2.0
